Rename HORSES_DATA_RL.ipynb to HORSES_DATA Reinforcement Learning.ipynb
Creating a Reinforcement Learning agent to pick horses to back/lay as appropriate. Backing multiple horses within a race. 

It will also decide on optimum bet sizes (as a fraction of current a/c size). This is tricky as it will need to account for the problem of ergodicity. This rears its head in systems with multiplicative dynamics. Usually handled by taking the log of the expected value (utility function), aka Kelly Criterion.